misc/HtmlTestSuite/results/test-0883.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>Interaction of 'inert' attribute with modal dialog, when the dialog is the root element</title>
<link href="mailto:obrufau@igalia.com" rel="author" title="Oriol Brufau"/>
<link href="https://html.spec.whatwg.org/multipage/interaction.html#inert" rel="help"/>
<meta content="Checks that being part of a modal dialog does not protect a node from being marked inert by an 'inert' attribute." name="assert"/>
</head><body><div id="log"/>
<dialog id="dialog">
  dialog
  <span id="child">
    child
  </span>
</dialog>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script>
const dialog = document.getElementById("dialog");
const root = document.documentElement;

setup(() =&gt; {
  root.remove();
  document.append(dialog);
  dialog.showModal();
  add_completion_callback(() =&gt; {
    getSelection().removeAllRanges();
  });
});

function checkSelection(expectedText) {
  const selection = getSelection();
  selection.selectAllChildren(document.documentElement);
  const actualText = selection.toString().trim();
  assert_equals(actualText, expectedText);
}

test(function() {
  checkSelection("dialog child");
}, "Modal dialog only marks outside nodes as inert");

test(function() {
  child.inert = true;
  this.add_cleanup(() =&gt; { child.inert = false; });
  checkSelection("dialog");
}, "Inner nodes with 'inert' attribute become inert anyways");

test(function() {
  dialog.inert = true;
  this.add_cleanup(() =&gt; { dialog.inert = false; });
  checkSelection("");
}, "If the modal dialog has the 'inert' attribute, everything becomes inert");

// Ideally this would happen in a completion callback, but then it would
// be too late: the results would have been shown inside the dialog.
dialog.remove();
document.append(root);
</script>
</body></html>
